# docs/swagger.json
{
    "openapi": "3.1.1",
    "info": {
        "title": "API para o Departamento de Polícia",
        "summary": "API para gerenciar casos e agentes policiais",
        "description": "Essa API permite o gerenciamento de casos policiais e agentes, incluindo operações CRUD para ambos.",
        "termsOfService": "http://localhost:3000/terms/",
        "contact": {
            "name": "Gabriel",
            "email": "gabriellvbs14@gmail.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Ambiente de desenvolvimento"
        }
    ],
    "paths": {
        "/agentes":{
            "get": {
                "summary": "Retorna uma lista de agentes",
                "description": "Essa rota retorna uma lista de todos os agentes cadastrados.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Id para filtrar agentes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form",
                        "explode": true
                    },
                    {
                        "name": "nome",
                        "in": "query",
                        "description": "Nome para filtrar agentes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form",
                        "explode": true
                    },
                    {
                        "name": "dataDeIncorporacao",
                        "in": "query",
                        "description": "dataDeIncorporacao para filtrar agentes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form",
                        "explode": true
                    },
                    {
                        "name": "cargo",
                        "in": "query",
                        "description": "Cargo para filtrar agentes",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form",
                        "explode": true
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Ordenar por um campo específico",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "style": "form",
                        "explode": true
                    }
                ],
                "tags": ["Gets"],
                "responses": {
                    "200": {
                        "description": "Retorna todos os agentes cadastrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "example": [
                                        {
                                            "id": "4dcd8f2a-1a2f-4786-af0a-d7baee70f270",
                                            "nome": "Rommel Carneiro",
                                            "dataDeIncorporacao": "1992-10-04",
                                            "cargo": "delegado"
                                        },
                                        {
                                            "id": "c93dea02-c8dd-4fa6-bba5-2bc2661dbff8",
                                            "nome": "Ana Silva",
                                            "dataDeIncorporacao": "2000-05-12",
                                            "cargo": "investigador"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Cria o registro de um novo agente",
                "description": "Essa rota cria o registro de um novo agente e retorna o agente criado.",
                "tags": ["Posts"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgenteRequestBody"
                            }
                        }
                    },
                    "required": ["nome", "dataDeIncorporacao", "cargo"]
                },
                "responses": {
                    "201": {
                        "description": "Retorna o agente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteSucesso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao criar o agente, devido a dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteErroParametros"
                                }
                            }
                        }
                    }
                }
            }
            
        },
        "/agentes/{id}":{
            "get": {
                "summary": "Retorna um agentes",
                "description": "Essa rota retorna um agente cadastrado com determinado id.",
                "parameters": [
                    {
                    "name": "id",
                    "in": "path",
                    "description": "Id do agente para ser retornado",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                    }
                ],
                "tags": ["Gets"],
                "responses": {
                    "200": {
                        "description": "Retorna o agente cadastrado com id especificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteSucesso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro ao encontrar o agente inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteInexistenteErro"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Atualiza completamente o registro de um agente",
                "description": "Essa rota atualiza completamente o registro de um agente e retorna o agente atualizado.",
                "parameters": [
                    {
                    "name": "id",
                    "in": "path",
                    "description": "Id do agente para ser atualizado",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                    }
                ],
                "tags": ["Puts"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgenteRequestBody"
                            }
                        }
                    },
                    "required": ["nome", "dataDeIncorporacao", "cargo"]
                },
                "responses": {
                    "200": {
                        "description": "Retorna o agente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteSucesso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar o agente, devido a dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteErroParametros"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro ao atualizar o agente inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteInexistenteErro"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "summary": "Atualiza parcialmente o registro de um agente",
                "description": "Essa rota atualiza parcialmente o registro de um agente e retorna o agente atualizado.",
                "parameters": [
                    {
                    "name": "id",
                    "in": "path",
                    "description": "Id do agente para ser atualizado",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                    }
                ],
                "tags": ["Patches"],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AgenteRequestBodyPatch"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o agente criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteSucesso"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erro ao atualizar parcialmente o agente, devido a dados inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteErroParametros"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro ao atualizar parcialmente o agente inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteInexistenteErro"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Deleta o registro de um agentes",
                "description": "Essa rota deleta o registro de um agente",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do agente para ser excluído",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "tags": ["Deletes"],
                "responses": {
                    "204": {
                        "description": "Agente excluido com sucesso, sem retorno"
                    },
                    "404": {
                        "description": "Erro ao deletar o agente inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteInexistenteErro"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/casos/{caso_id}/agente":{
            "get":{
                "summary": "Retorna o agente responsável pelo caso",
                "description": "Essa rota retorna os dados do agente responsável pelo caso cadastrado com determinado id.",
                "parameters": [
                    {
                    "name": "caso_id",
                    "in": "path",
                    "description": "Id do caso para retornar o agente responsável",
                    "required": true,
                    "schema": {
                        "type": "string",
                        "format": "uuid"
                    }
                    }
                ],
                "tags": ["Gets"],
                "responses": {
                    "200": {
                        "description": "Retorna os dados do agente responsável pelo caso cadastrado com id especificado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AgenteSucesso"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Erro ao buscar o caso inexistente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CasoInexistenteErro"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components":{
        "schemas":{
            "AgenteRequestBody": {
                "type": "object",
                "required": ["nome", "dataDeIncorporacao", "cargo"],
                "properties":{
                    "nome": {
                        "type": "string"
                    },
                    "dataDeIncorporacao": {
                        "type": "string",
                        "format": "date"
                    },
                    "cargo": {
                        "type": "string"
                    }
                }
            },
            "AgenteRequestBodyPatch": {
                "type": "object",
                "required": [],
                "properties":{
                    "nome": {
                        "type": "string"
                    },
                    "dataDeIncorporacao": {
                        "type": "string"
                    },
                    "cargo": {
                        "type": "string"
                    }
                }
            },
            "AgenteErroParametros": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 400
                    },
                    "message": {
                        "type": "string",
                        "example": "Parâmetros inválidos"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "nome": "A requisição deve possuir o campo 'nome'",
                            "dataDeIncorporacao": "A requisição deve possuir o campo 'dataDeIncorporacao'",
                            "cargo": "A requisição deve possuir o campo 'cargo'"
                        }
                    }
                }
            },
            "AgenteInexistenteErro": {
                "type": "object",
                "properties":{
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "type": "string",
                        "example": "Agente inexistente"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "id": "Não existe agente com esse id"
                        }
                    }
                }
            },
            "AgenteSucesso": {
                "type": "object",
                "required": ["id", "nome", "dataDeIncorporacao", "cargo"],
                "properties":{
                    "id":{
                        "type": "string",
                        "format": "uuid"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "dataDeIncorporacao": {
                        "type": "string",
                        "format": "date"
                    },
                    "cargo": {
                        "type": "string"
                    }
                }
            },
...
            "CasoInexistenteErro": {
                "type": "object",
                "properties":{
                    "status": {
                        "type": "integer",
                        "example": 404
                    },
                    "message": {
                        "type": "string",
                        "example": "Caso inexistente"
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "id": "Não existe caso com esse id"
                        }
                    }
                }
            },
...
        }
    }
}